Homework #7. Network.

+++ app/src/main/java/ru/asshands/softwire/androidacademy2020/data/Movie.kt
@@ -5,13 +5,13 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class Movie(
-    val id: Int,
+    val id: Long,
     val title: String,
     val overview: String,
     val poster: String,
     val backdrop: String,
-    val ratings: Float,
-    val numberOfRatings: Int,
+    val ratings: Double,
+    val numberOfRatings: Long,
     val minimumAge: Int,
     val runtime: Int,
     val genres: List<Genre>,
